Move `ContractRunner` interface to `contract/types.ts`
